requirements1.txt
streamlit
python-dotenv
langchain
unstructured==0.7.12
unstructured[pdf]
openai
chromadb
fitz
tiktoken